+++ core/src/main/scala/org/locationtech/rasterframes/expressions/package.scala
@@ -93,6 +93,8 @@ package object expressions {
     registry.registerExpression[Min]("rf_local_min")
     registry.registerExpression[Max]("rf_local_max")
     registry.registerExpression[Clip]("rf_local_clip")
+    registry.registerExpression[Where]("rf_where")
+    registry.registerExpression[Standardize]("rf_standardize")
     registry.registerExpression[Sum]("rf_tile_sum")
     registry.registerExpression[Round]("rf_round")
     registry.registerExpression[Abs]("rf_abs")